move fetchAndSaveItems method to harvester_service from harvester_event_worker, and add integration test
